Extract shared integration test server and document required requests
